algebra: new data types
Added a bunch of new data types following the GLSL standard data types. All for C++ language.

+++ type/dvec2.hpp
@@ -5,13 +5,13 @@ using vec2 = union u_vec2;
 
 union u_vec2 {
     struct {
-        float x;
-        float y;
+        double x;
+        double y;
     };
 
     struct {
-        float w;
-        float h;
+        double w;
+        double h;
     };
 };
